rename config to userconfig
to differentiate between a user configuration file, and an application
configuration resource
