+++ cmd/list.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/joakimen/kf/pkg/config"
 	"github.com/joakimen/kf/pkg/slice"
+	"github.com/joakimen/kf/pkg/userconfig"
 	"github.com/spf13/cobra"
 )
 
@@ -13,7 +13,7 @@ var listCmd = &cobra.Command{
 	Use:   "list",
 	Short: "List all known files",
 	Run: func(cmd *cobra.Command, args []string) {
-		configFileLines, err := config.ReadConfigFile()
+		configFileLines, err := userconfig.ReadConfigFile()
 		if err != nil {
 			fmt.Fprintf(os.Stderr, "error reading configuration file: %v\n", err)
 			os.Exit(1)
